+++ package.json
@@ -53,7 +53,7 @@
     "expo-constants": "~16.0.2",
     "expo-device": "~6.0.2",
     "expo-file-system": "~17.0.1",
-    "expo-font": "~12.0.9",
+    "expo-font": "~12.0.10",
     "expo-haptics": "^13.0.1",
     "expo-image-manipulator": "~12.0.5",
     "expo-image-picker": "~15.0.7",
