+++ src/services/retired/rename_images.py
